# Conflicts:
#	DataConverter.Test/ShipTests/WgShipDeserialization.cs
#	WoWsShipBuilderDataStructures/EnumUtils.cs

+++ DataConverter/WGStructure/WGAircraft.cs
@@ -8,7 +8,7 @@ namespace DataConverter.WGStructure
     {
         public double attackCooldown { get; set; }
         public int attackCount { get; set; }
-        public int attackInterval { get; set; }
+        public double attackInterval { get; set; }
         public double attackSpeedMultiplier { get; set; }
         public double attackSpeedMultiplierApplyTime { get; set; }
         public double attackerDamageTakenMultiplier { get; set; }
@@ -28,7 +28,7 @@ namespace DataConverter.WGStructure
         public bool isJatoBoosterDetachable { get; set; }
         public double jatoDuration { get; set; }
         public double jatoSpeedMultiplier { get; set; }
-        public int maxHealth { get; set; }
+        public float maxHealth { get; set; }
         public string name { get; set; }
         public double naturalAcceleration { get; set; }
         public double naturalDeceleration { get; set; }
